feat: Use a thread pool in order to create multiple EventLoopThread(s) (#32)
When we cache non async functions we use a single EventLoopThread to
execute it since all of the machinery of GCache is async otherwise.

However, this presents issues because if we have multiple requests they
will only execute in single thread, and since we are executing sync
functions they will block each other.

With this change, we will be using a simple wrapper on top of
EventLoopThread which manages multiples of them and then picks one at
random to execute a coroutine.